ref(normalization): Split validation and normalization
