+++ relay-server/src/metrics_extraction/event.rs
@@ -63,7 +63,10 @@ pub fn extract_metrics(event: &Event, config: &MetricExtractionConfig) -> Vec<Bu
 mod tests {
     use chrono::{DateTime, Utc};
     use relay_dynamic_config::{Feature, FeatureSet, ProjectConfig};
-    use relay_event_normalization::{normalize_event, NormalizationConfig};
+    use relay_event_normalization::{
+        normalize_event, validate_event_timestamps, validate_transaction, EventValidationConfig,
+        NormalizationConfig, TransactionValidationConfig,
+    };
     use relay_event_schema::protocol::Timestamp;
     use relay_protocol::Annotated;
     use std::collections::BTreeSet;
@@ -497,7 +500,13 @@ mod tests {
         let mut event = Annotated::from_json(json).unwrap();
         let features = FeatureSet(BTreeSet::from([Feature::SpanMetricsExtraction]));
 
-        // Normalize first, to make sure that all things are correct as in the real pipeline:
+        // Validate and normalize first, to make sure that all things are correct as in the real pipeline:
+        let res = validate_transaction(&event, &TransactionValidationConfig::default());
+        assert!(res.is_ok());
+
+        let res = validate_event_timestamps(&mut event, &EventValidationConfig::default());
+        assert!(res.is_ok());
+
         normalize_event(
             &mut event,
             &NormalizationConfig {
@@ -505,8 +514,7 @@ mod tests {
                 normalize_spans: true,
                 ..Default::default()
             },
-        )
-        .unwrap();
+        );
 
         // Create a project config with the relevant feature flag. Sanitize to fill defaults.
         let mut project = ProjectConfig {
@@ -1027,8 +1035,7 @@ mod tests {
                 normalize_spans: true,
                 ..Default::default()
             },
-        )
-        .unwrap();
+        );
 
         // Create a project config with the relevant feature flag. Sanitize to fill defaults.
         let mut project = ProjectConfig {
@@ -1141,8 +1148,7 @@ mod tests {
                 device_class_synthesis_config: true,
                 ..Default::default()
             },
-        )
-        .unwrap();
+        );
 
         // Create a project config with the relevant feature flag. Sanitize to fill defaults.
         let mut project = ProjectConfig {
