+++ relay-server/tests/test_fixtures.rs
@@ -1,7 +1,8 @@
 use std::fs;
 
 use relay_event_normalization::{
-    normalize_event, NormalizationConfig, StoreConfig, StoreProcessor,
+    normalize_event, validate_event_timestamps, validate_transaction, EventValidationConfig,
+    NormalizationConfig, StoreConfig, StoreProcessor, TransactionValidationConfig,
 };
 use relay_event_schema::processor::{process_value, ProcessingState};
 use relay_event_schema::protocol::Event;
@@ -71,7 +72,9 @@ macro_rules! event_snapshot {
             fn test_processing() {
                 let mut event = load_fixture();
 
-                normalize_event(&mut event, &NormalizationConfig::default()).unwrap();
+                validate_transaction(&event, &TransactionValidationConfig::default()).unwrap();
+                validate_event_timestamps(&mut event, &EventValidationConfig::default()).unwrap();
+                normalize_event(&mut event, &NormalizationConfig::default());
 
                 let config = StoreConfig::default();
                 let mut processor = StoreProcessor::new(config, None);
