+++ tools/process-event/src/main.rs
@@ -10,7 +10,8 @@ use std::path::PathBuf;
 use anyhow::{format_err, Context, Result};
 use clap::Parser;
 use relay_event_normalization::{
-    normalize_event, NormalizationConfig, StoreConfig, StoreProcessor,
+    normalize_event, validate_event_timestamps, validate_transaction, EventValidationConfig,
+    NormalizationConfig, StoreConfig, StoreProcessor, TransactionValidationConfig,
 };
 use relay_event_schema::processor::{process_value, ProcessingState};
 use relay_event_schema::protocol::Event;
@@ -83,8 +84,11 @@ impl Cli {
         }
 
         if self.store {
-            normalize_event(&mut event, &NormalizationConfig::default())
+            validate_transaction(&event, &TransactionValidationConfig::default())
                 .map_err(|e| format_err!("{e}"))?;
+            validate_event_timestamps(&mut event, &EventValidationConfig::default())
+                .map_err(|e| format_err!("{e}"))?;
+            normalize_event(&mut event, &NormalizationConfig::default());
             let mut processor = StoreProcessor::new(StoreConfig::default(), None);
             process_value(&mut event, &mut processor, ProcessingState::root())
                 .map_err(|e| format_err!("{e}"))
